[gcc] Scene: Make float to int conversion explicit

+++ src/scene/Scene.cpp
@@ -1213,8 +1213,8 @@ static void ARX_PORTALS_Frustrum_RenderRoomTCullSoft(size_t room_num,
 					fr=std::min(fr,255.f);
 					fb*=255.f;
 					fb=std::min(fb,255.f);
-					u8 lfr = fr;
-					u8 lfb = fb;
+					u8 lfr = u8(fr);
+					u8 lfb = u8(fb);
 					u8 lfg = 0x1E;
 					
 					ep->color[k] = Color(lfr, lfg, lfb, 255).toRGBA();
